requirements.txt
Django==1.8.3
django-braces==1.8.1
django-cors-headers==1.1.0
django-oauth-toolkit==0.8.2
django-phonenumber-field==0.7.2
django-tastypie==0.12.2
django-tastypie-oauth==0.0.3
oauthlib==1.0.0
phonenumbers==7.0.8
Pillow==2.9.0
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
